+++ src/main/java/com/saintdan/framework/constant/CommonsConstant.java
@@ -9,17 +9,15 @@ package com.saintdan.framework.constant;
  */
 public interface CommonsConstant {
 
-  String PARAM_BLANK = "Param %s cannot be null.";
   String ID_BLANK = "Param id cannot be null.";
   String ID = "id";
   String NAME = "name";
   String CREATED_BY = "createdBy";
   String LAST_MODIFIED_BY = "lastModifiedBy";
-  String LAST_MODIFIED_DATE = "lastModifiedDate";
   String VALID_FLAG = "validFlag";
   String BLANK = "";
   String COMMA = ",";
   String COLON = ":";
-  String SLASH = "/";
+  String UNDERLINE = "_";
   String ZERO = "0";
 }