+++ src/main/java/com/saintdan/framework/controller/app/HeartbeatController.java
@@ -1,9 +1,10 @@
 package com.saintdan.framework.controller.app;
 
-import com.saintdan.framework.constant.ResourceURL;
+import com.saintdan.framework.constant.ResourcePath;
 import com.saintdan.framework.constant.VersionConstant;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
+import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -19,13 +20,15 @@ import org.springframework.web.bind.annotation.RestController;
  */
 @Api(value = "Heartbeat")
 @RestController
-@RequestMapping(
-    ResourceURL.RESOURCES + VersionConstant.V1 + ResourceURL.APP + ResourceURL.HEARTBEAT)
+@RequestMapping(ResourcePath.API + VersionConstant.V1 + ResourcePath.APP + ResourcePath.HEARTBEAT)
 public class HeartbeatController {
 
   @RequestMapping(method = RequestMethod.GET)
   @ApiOperation(value = "Heartbeat", httpMethod = "GET", response = ResponseEntity.class)
-  @ApiImplicitParam(name = "Authorization", value = "token", paramType = "header", dataType = "string", required = true)
+  @ApiImplicitParams({
+      @ApiImplicitParam(name = "Authorization", value = "token", paramType = "header", dataType = "string", required = true),
+      @ApiImplicitParam(name = "Limit-Key", value = "limit key", paramType = "header", dataType = "string")
+  })
   public ResponseEntity heartbeat() {
     return ResponseEntity.ok().build();
   }