+++ src/main/java/com/saintdan/framework/controller/open/LoginController.java
@@ -1,17 +1,15 @@
 package com.saintdan.framework.controller.open;
 
 import com.saintdan.framework.component.ResultHelper;
-import com.saintdan.framework.component.ValidateHelper;
-import com.saintdan.framework.constant.ResourceURL;
+import com.saintdan.framework.constant.ResourcePath;
 import com.saintdan.framework.constant.VersionConstant;
 import com.saintdan.framework.enums.ErrorType;
-import com.saintdan.framework.enums.GrantType;
-import com.saintdan.framework.exception.IllegalTokenTypeException;
 import com.saintdan.framework.param.LoginParam;
 import com.saintdan.framework.service.LoginService;
 import com.saintdan.framework.tools.Assert;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
+import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
 import javax.servlet.http.HttpServletRequest;
 import org.slf4j.Logger;
@@ -28,24 +26,18 @@ import org.springframework.web.bind.annotation.RestController;
 
 @Api("Login")
 @RestController
-@RequestMapping(ResourceURL.RESOURCES + VersionConstant.V1 + ResourceURL.OPEN + ResourceURL.LOGIN)
+@RequestMapping(
+    ResourcePath.API + VersionConstant.V1 + ResourcePath.OPEN + ResourcePath.LOGIN)
 public class LoginController {
 
   @RequestMapping(method = RequestMethod.POST)
   @ApiOperation(value = "Login", httpMethod = "POST", response = OAuth2AccessToken.class)
-  @ApiImplicitParam(name = "Authorization", value = "token", paramType = "header", dataType = "string", required = true)
-  public ResponseEntity postAccessToken(HttpServletRequest request, @RequestBody LoginParam param)
+  @ApiImplicitParams({
+      @ApiImplicitParam(name = "Authorization", value = "token", paramType = "header", dataType = "string", required = true),
+      @ApiImplicitParam(name = "Limit-Key", value = "limit key", paramType = "header", dataType = "string")
+  })
+  public ResponseEntity login(HttpServletRequest request, @RequestBody LoginParam param)
       throws HttpRequestMethodNotSupportedException {
-    // Validate current user, param and sign.
-    ResponseEntity responseEntity;
-    try {
-      responseEntity = validateHelper.validate(request, param, GrantType.PASSWORD);
-    } catch (IllegalTokenTypeException e) {
-      return resultHelper.infoResp(e.getErrorType(), HttpStatus.UNPROCESSABLE_ENTITY);
-    }
-    if (!responseEntity.getStatusCode().is2xxSuccessful()) {
-      return responseEntity;
-    }
     try {
       return loginService.login(param, request);
     } catch (Exception e) {
@@ -58,16 +50,12 @@ public class LoginController {
   private static final Logger logger = LoggerFactory.getLogger(LoginController.class);
   private final LoginService loginService;
   private final ResultHelper resultHelper;
-  private final ValidateHelper validateHelper;
 
-  @Autowired public LoginController(LoginService loginService, ResultHelper resultHelper,
-      ValidateHelper validateHelper) {
+  @Autowired public LoginController(LoginService loginService, ResultHelper resultHelper) {
     Assert.defaultNotNull(loginService);
     Assert.defaultNotNull(resultHelper);
-    Assert.defaultNotNull(validateHelper);
     this.loginService = loginService;
     this.resultHelper = resultHelper;
-    this.validateHelper = validateHelper;
   }
 }