+++ src/main/java/com/saintdan/framework/param/ClientParam.java
@@ -2,13 +2,12 @@ package com.saintdan.framework.param;
 
 import com.saintdan.framework.annotation.NotNullField;
 import com.saintdan.framework.domain.ClientDomain;
-import com.saintdan.framework.enums.OperationType;
-import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
+import org.springframework.http.HttpMethod;
 
 /**
  * Param bean for {@link ClientDomain}
@@ -26,21 +25,20 @@ public class ClientParam extends BaseParam {
 
   private static final long serialVersionUID = 6065608866944007796L;
 
-  @NotNullField(value = {OperationType.UPDATE,
-      OperationType.DELETE}, message = "id cannot be null.")
+  @NotNullField(method = {HttpMethod.PUT, HttpMethod.DELETE}, message = "id cannot be null.")
   private Long id;
 
-  @ApiModelProperty(notes = "Name of service.")
+  @NotNullField(method = HttpMethod.POST, message = "name cannot be null.")
   private String name;
 
-  @NotNullField(value = OperationType.UPDATE, message = "publicKey cannot be null.")
+  @NotNullField(method = HttpMethod.POST, message = "publicKey cannot be null.")
   private String publicKey;
+
+  @NotNullField(method = HttpMethod.POST, message = "scope cannot be null.")
   private String scope;
+
+  @NotNullField(method = HttpMethod.POST, message = "grantType cannot be null.")
   private String grantType;
   private Integer accessTokenValiditySeconds;
   private Integer refreshTokenValiditySeconds;
-
-  public ClientParam(Long id) {
-    this.id = id;
-  }
 }